Generalizing stubs file
